requirements.txt
regex
numpy
tensorflow
transformers
customtkinter
pillow